trans-bridge/requirements-bridge.txt
minio==7.2.15
asyncio
uuid
aiokafka
google-cloud-run